import/requirements.txt
PyYAML==6.0.0
